fix: do not export as a constant but a function
If you export configs as a constant, some configurations refer dependencies specific to its config on runtime (when running `eslint .`).
For example, eslint-config-next refers files in `next` package even your eslint.config.js does not call `next` rule of this package. If `next` is not installed, error raises and the lint is not executed.
To avoid this, I export functions to return configs so that the dependencies like `next` is not called on running `eslint` command until the config for next is called.