Reason why I abandoned this branch
I considered to merge ESLint configs and tsconfigs to one package, but I found tsconfig is required on the build process while ESLint is not.
If I merge ESLint configs and tsconfigs, you have to a lot of eslint-configs on build time and it slows down the build process.

I keep this branch as a backup because it might be useful when I add other linters (e.g. stylelint, markuplint, etc.) to the ESLint package.